tests/wpt/web-platform-tests/css/css-box/parsing/padding-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: padding sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-padding">
<meta name="assert" content="padding supports the full grammar '<length-percentage>{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('padding', '1px 2px 3px 4px', {
  'padding-top': '1px',
  'padding-right': '2px',
  'padding-bottom': '3px',
  'padding-left': '4px'
});

test_shorthand_value('padding', '1px 2px 3px', {
  'padding-top': '1px',
  'padding-right': '2px',
  'padding-bottom': '3px',
  'padding-left': '2px'
});

test_shorthand_value('padding', '1px 2px', {
  'padding-top': '1px',
  'padding-right': '2px',
  'padding-bottom': '1px',
  'padding-left': '2px'
});

test_shorthand_value('padding', '1px', {
  'padding-top': '1px',
  'padding-right': '1px',
  'padding-bottom': '1px',
  'padding-left': '1px'
});
</script>
</body>
</html>
